make it avaliable to major package managers and switch to curl

+++ install.sh
@@ -1,21 +1,49 @@
-# check if wget & make installed first
-sudo apt update
-sudo apt install bzip2 make wget -y
+#!/bin/bash
 
-# download pre-built btop files
-wget -O btop.tbz https://github.com/aristocratos/btop/releases/latest/download/btop-x86_64-linux-musl.tbz
+package_manager=''
+package_managers=("yum" "dnf" "pacman" "apt" "nix")
 
-# create a directory and decompress
+declare -A package_manager_commands
+package_manager_commands=(
+  ["yum"]="yum install -y"
+  ["dnf"]="dnf install -y"
+  ["pacman"]="pacman --noconfirm -S"
+  ["apt"]="apt install -y"
+  ["nix"]="nix-env -i"
+)
+
+for i in "${package_managers[@]}"
+do
+  if which $i &> /dev/null; then
+    package_manager=$i
+    break
+  fi
+done
+
+if [ -z $package_manager ]; then
+  echo "No package manager found"
+  exit 1
+else
+  command=${package_manager_commands[$package_manager]}
+fi
+
+sudo $command btop
+
+if [ $? -eq 0 ]; then
+  exit 0
+fi
+
+sudo $command bzip2 make curl
+
+curl -Lo btop.tbz https://github.com/aristocratos/btop/releases/latest/download/btop-x86_64-linux-musl.tbz
+
 mkdir btop
 tar -xjf btop.tbz -C btop
 
-# remove file
 rm btop.tbz
 
-# switch to btop directory
 cd btop/btop
 
-# run install script
 ./install.sh
 ./setuid.sh
